--- src/redux/configStore.js
@@ -1,7 +1,0 @@
-import {createStore, combineReducers} from "redux";
-import post from "./modules/post";
-
-const rootReducer = combineReducers({post});
-const store = createStore(rootReducer);
-
-export default store;
